Update patch set 2
Patch Set 2:

> Making sure the output is in the correct range is what this patch is about, not the precision (in fact against precision in the latest version) or optimization. The sample code exhibits the very same bug I'm trying to squash here. Test with gain = 1.0f and someValue anything between 2147483584 and 2147483647. Or whatever values that should give output in that range for the gain you choose.

Yes, but what you did introduce other problems. The range is now off-center (because it does not go all the way to int32_max) so the midpoint will not be at 0. And all values will be slightly modified even if they are well into the acceptable range.

I think what we should do is proper clamping: convert the float to the range int32_min to int32_max, and if the conversion ends up being out of range, force it to be int32_max (or min, if it is out of range on the other side). This will have no consequences on the remaining range, only the loss of precision from the float conversion.

Patch-set: 2