Update patch set 6
Patch Set 6: Code-Review-1

It seems to me like a lot of the compat_ structures are simply not needed, or could be made not needed. E.g. timespec and user_timer_info consist only of fixed-size integers, and so as long as we specify _PACKED, or a fixed alignment, they will be the same size everywhere.

In fact, since 64-bit is or will be the dominant platform going forward (ARM64, etc.), it may just make sense to have all syscall structure arguments be 64-bit, and "waste" the extra space for pointers on 32-bit. That will allow us to get rid of the copy_var_ref... altogether, as we can then just add STATIC_ASSERT(sizeof(struct) == N) to validate that the structures are the same size everywhere.

Patch-set: 6
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-1