Update patch set 1
Patch Set 1:

> We should probably copy whatever Linux/BSD/etc. do on SPARC regarding where the kernel lives in virtual memory.

As I said in the commit message, on sparc the kernel has an independant address space, so we can't use "pointer tagging" as we do on x86 to decide if an address is kernel or userland. The same address can be valid in both kernel and userspage, pointing to different things. We'll need a separate boolean instead in places where both kernel and user addresses can be used.

Anyway, I can revisit this when we have some userland up and running.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>