sparc: move stack ouf of the way
This was copied from PPC and is not done on any other architecture.

Having the stack use the kernel address results in the kernel being
loaded elsewhere. Not really a problem (the kernel is relocatable), but
it's annoying when debugging as addresses need to be converted between
the elf file and the actual loading address.