Update patch set 7
Patch Set 7:

Comment from Axel 8 years ago on Trac: "I was actually aware of that issue. However, PPC allows unaligned access as well, so I thought it would work out okay.

Wouldn't it be possible to allow unaligned access in the getter/setter methods if the architecture allows it? While that would make testing harder, it might also be worth the hassle for the extra performance gain."

Patch-set: 7
CC: Gerrit User 1000023 <1000023@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>