Update patch set 4
Patch Set 4:

> Does it do the proper save/restore now?

Yes, and it even manages to return cleanly to the bootloader if the kernel entry function exits (which somehow happened accidentally before one of the other changes). The bootloader doesn't really do anything useful in that case however, it just prints an error message and stops. But it's better than crashing :)

Patch-set: 4