Update patch set 1
Patch Set 1:

One idea how this situation can form: The package server is the entity through which effective package state is changed (i.e. transactions are committed in order, one after the other). There may be multiple clients using the Package Kit, which creates *concurrent* transactions. For example installing software in HaikuDepot uses asynchronous transactions, or when you use pkgman in parallel in two or more Terminals. Then it may happen that the same package is installed by more than one transaction, if it was missing when each them started. The transactions are however not aware of each other. One of the transactions is committed first, then the package file is already present at the time when the second transaction is to be committed.

Patch-set: 1
Reviewer: Gerrit User 1000024 <1000024@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>