Update patch set 2
Patch Set 2:

> Patch Set 1:
> One idea how this situation can form: The package server is the entity through which effective package state is changed (i.e. transactions are committed in order, one after the other). There may be multiple clients using the Package Kit, which creates *concurrent* transactions. For example installing software in HaikuDepot uses asynchronous transactions, or when you use pkgman in parallel in two or more Terminals. Then it may happen that the same package is installed by more than one transaction, if it was missing when each them started. The transactions are however not aware of each other. One of the transactions is committed first, then the package file is already present at the time when the second transaction is to be committed.

In such a situation the package placed there by another transaction would be activated and clobbering an in-use package is probably a bad idea. Or is this handled gracefully by producing a package de- and reactivation?

As all transactions go through the single package_daemon, why don't we simply synchronize them so that there cannot be any concurrent transactions? In the best case it is wasted bandwidth to download common packages twice.

If there is a package for other reasons than concurrent transactions, the new download could be skipped as well. How this should work probably depends on whether the new package is identical to the present one or not. If it is identical (which should be easy to check based on its checksum), the download can be skipped, if it is not, then it could clobber and possible de- and reactivate.

Patch-set: 2
CC: Gerrit User 1000017 <1000017@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>