Package Server: clobber files when commiting transaction
Package files to be committed may already exist in the packages folder,
there is no good reason to fail the transaction. These may be packages
which are not activated, or there may be other reasons. While I have
not investigated how this situation may form, I don't see a good reason
to spend the time. Simply move the BEntry with "clobber = true".

Fixes #15068.

Change-Id: I2a574df38385c8dead8bac2beace94436d43760a