Update patch set 2
Patch Set 2:

> If there is a package for other reasons than concurrent transactions, the new download could be skipped as well. How this should work probably depends on whether the new package is identical to the present one or not. If it is identical (which should be easy to check based on its checksum), the download can be skipped, if it is not, then it could clobber and possible de- and reactivate.

The package daemon and pkgman will already do this when deciding what to install. We get in this situation only if they decided that the existing package needs to be replaced. This can be because an user is force-installing it (pkgman install file.hpkg) in which case there is no download. I think this is the only case where I have hit this problem.

In this case, the old package is probably already scheduled to be deleted in the transaction, it just happens that the new file is copied in before the old one is erased.

There is a variant of this in haikuports where the packagefs is rebuilt from scratch everytime haikuporter is run. In this case, there are files already in the package dir but they are not activated, and a first transaction will try to replace them. In that case, there is also no download involved.

I am not aware of cases where a file is incorrectly re-downloaded. I guess it could indeed happen when running multiple transactions in parallel, and that could also be improved to save on download times, but that seems to be a separate issue, and this fix will still be useful regardless.

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>