Update patch set 2
Patch Set 2:

> In such a situation the package placed there by another transaction would be activated and clobbering an in-use package is probably a bad idea. Or is this handled gracefully by producing a package de- and reactivation?

I see your point, but I thought BEntry::Move() is an atomic operation on the FS level.

> As all transactions go through the single package_daemon, why don't we simply synchronize them so that there cannot be any concurrent transactions? In the best case it is wasted bandwidth to download common packages twice.

This would be quite a re-design of the Package Kit. I mean "what lives where". The issue is only with packages that multiple other packages depend on. After a while, such packages will tend to be already installed. So it should become less of an issue. I can see however, how it could be bad on a fresh system where one quickly installs multiple Qt based apps without waiting for one install to finish first.

The big question is whether it is OK to replace a file, even if atomically. The assumption here is that the file contents are identical, which I believe they should be, since the file name already makes a package file unique. It could of course be a wrong assumption if multiple repositories contain the "same" package, but they are actually slightly different. In any case, the solution could also be to not move the BEntry and ignore the error. Or check if the files are really equal and only then ignore the error.

Patch-set: 2
Reviewer: Gerrit User 1000024 <1000024@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>