Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Please have another log. Now the clobber move is avoided, and the
> two files are compared. Only fail if they are different (which
> validates the original assumption that clobbering would be OK).
> 
> I managed to trigger this code by manually moving all packages
> from the "_packages_" optional packages folder into "packages"
> and confirmed I saw the added output from this patch in the
> syslog.

Now I'm confused.

One of the cases where I frequently hit this problem is the following scenario:

- I create a package (for example with haikuporter)
- I install it and make some tests
- I find something is not right with my package
- I create another version of the package (different but with the same name)
- I use pkgman to install this new version (pkgman install mynewpackage.hpkg)
- pkgman tells me that the package will be updated with the new version (which has the same name)
- I say ok
- I get a "name in use" error

Your initial patch looked like it would fix that, but now the new version intentionally doesn't?

Compared to the following scenario:
- I install a package
- I try to install the same package again, without any change to it 
- This should be intercepted at a much higher level by pkgman, which should tell me there is nothing to do (it doesn't currently)
- I shouldn't even get the confirmation prompt for an update from pkgman in this case, and the code shouldn't go as far as attempting to install packages that are already there.

I'm ready to accept that both of these should not be allowed, and if I want to do this, I need to manually uninstall and reinstall the package. But in that case, I would expect that the error is handled at pkgman level, telling me that there is nothing to do because the package I'm trying to install is already installed. Not a confusing "name in use" error after it told me that it would do an update?

Patch-set: 3