Update patch set 5
Patch Set 5:

(1 comment)

So, I looked into why my described use case ends up in this situation.

The change is hrev51233 and the ticket it fixes is #10038

What was happening was the following:
- When updating a package with another one of the same name, the old package would first be deleted, then the new one added
- This means that during a short time, the package would be missing
- The package server would notice that, and offer to uninstall all its dependents, especially annoying when you do this, for example, with the Haiku package.

The merged fix is in the package daemon to filter out the package from the transaction. But as a result, the old file is never removed, and this prevent the new one from being installed.

So, it seems this situation has to be handled differently.

After looking at this and discussing this in more details with stippi, I agree that this version of the change is more correct. It will not fix the use case of replacing a package with another one of the same version, but the problem in that case is caused by the patch in hrev51233, which is not correct, and should be fixed at that level.

Patch-set: 5
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>