bfs: fix misaligned access
The BFS on-disk data is not aligned. Reading it to memory and trying to
access fields directly does not work on sparc. memcpy the data to an
aligned variable before handling it with its native size.

We need to tell gcc that the data is not aligned. This is done with the
"packed" attribute, but it works only on structs. So we have to wrap the
values in a struct. Thanks to C++ features, we can make the struct
relatively transparent by having an assigment operator (for writes) and
a cast operator (for read access). The rest of the code is then largely
unchanged (except for use in printf statements and other vararg
functions, where the implicit casting can't work).

gcc takes care of performing the access in the correct way on platforms
that need it (old ARM, sparc) and can still optimize things on other
architectures.

Fixes #9255

Change-Id: I3bf62590dee059ad32b1845bdc4eace165b73203