Update patch set 12
Patch Set 12:

> Patch Set 12:
> 
> (1 comment)
> 
> I'm too dumb for this solution :-)
> How is that supposed to work?

The __attribute__((packed)) is enough for gcc to know that the value in the struct may be misaligned, and it will use appropriate code to access it.

The operator= and cast operators in the struct avoid adding .value everywhere in the code when we access it, by providing implicit conversions that let the existing code syntax do the right thing. operator= is used for write access, and the cast operator for read access.

Then it is only a matter of using this type where needed. We even get a compiler error if we attempt to cast a pointer to the struct to a pointer to the normal integer type.

I have confirmed that this version creates no regression for the sparc bootloader.

Patch-set: 12
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>