Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Why does the commit message say "fix cache size"? What was the problem observed with "3 x 1024 x 1024" cache size before?

The previous code was not checking for IsVideo() and was happily using colorspace and bytes per row of the audio stream to compute a size. I don't know what the result was.

We can decide to keep 3MB for audio streams, but the code wasn't doing that.

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>