Update patch set 7
Patch Set 7:

>  * DataReceived() gets an offset and uses BFile::WriteAt(). I have to see what the values actually are for range-requests and why WriteAt() is used in the first place.

WriteAt is used to handle redirections. In the case of a redirection (if the automatic redirection code in the network API is used), first you get data from the initial page, and then you get data from the redirected page, with the position starting back at 0.

In this case, it should start at the correct position, but that part is implemented in the network kit yet. It would need to extract the position from the HTTP headers in the resonse. Looking into it, I see that it isn't even completely correct without range requests, there is a problem if a redirection between two pages using gzip compression.

Designing the code this way makes it usable for other protocols that may send the data completely out of order (for example, bitttorrent, where incoming data may be from multiple peers, and they will each send separate parts of the file), or in the past it was used for "download accelerators" where a server would limit the maximum speed of a single TCP connection, and it was thus faster to do 4 range requests downloading 4 parts of the file, instead of one downloading the whole file.

In these situations, the WriteAt always writes at the correct position (or at least it's supposed to).

Note that during this summer, Leorize worked on new network APIs that remove this callback system and instead the network kit gets passed a DataIO where it can put the data directly. However this has not been integrated in Haiku yet, because it breaks the ABI and would break existing applications. Sadly, the more work we do on improving the current code, the harder it will be to merge this work.

Patch-set: 7
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>