Update patch set 1
Patch Set 1: Code-Review-1

I do not think we should add this code here. Since usb_disk obviously does not use isochronous requests, this implies that endpoint->max_burst_payload == 0, which probably means that the max_packet_size was found to be 0. I did not look up the specification, but that may mean we are supposed to use some kind of default (or that the device itself is broken.) At the very least we should add the check there and refuse to initialize the endpoint.

Patch-set: 1
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-1