Update patch set 7
Patch Set 7:

> Thanks for the explanation! I am not sure though -- can I now use range-requests considering that WriteAt() is supposed to be used? Or will that have to be accounted for in the request implementation?

We could workaround the problem in the application side (applying an offset matching the range request start to the WriteAt call) but it seems better to fix the problem in the network kit code instead.

What we will need is:
- Add a SetRange() call to BHttpRequest, to be called before starting the request
- That would set the appropriate header, and also store the start offset somewhere in the BHttpRequest (I think the needed fields are already there as this had been planned but the code never written)
- When calling the DataReceived callback, apply that offset to the parameter, so that applications can handle it easily (they will just write each chunk directly where it's supposed to go in the file)

Then all we will need in this side is:
- Call SetRange with a start at the current downloaded filesize, and end at -1 (which means end of the file on the server)
- Do not truncate the file in this case

I don't know if I will have time to look into the network part in the next few days. I think we can merge this version of the patch and keep this for another enhancement ticket?

Patch-set: 7
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>