BHttpRequest: Treat incomplete downloads as error
When BSocket::Read() returns 0, it was treated as
a finished transfer. This is OK when we don't know
the content length, but when we do, there is no reason
not to bubble up the error. Return B_IO_ERROR in this
case.

Also change the error mapping in BSecureSocket to
use B_IO_ERROR for SSL_ERROR_ZERO_RETURN.
Otherwise no retries happen in Package Kit downloads,
as B_CANCELED is not one of the errors considered
retryable, and I am not sure it should be.

Change-Id: I68801dbbb85bcfd2e7aa68fd6a9ded6304e82b19