Update patch set 1
Patch Set 1: Code-Review-1

Please don't send this to beta, without at least increasing the audio magic number: #16738.

After some logging in ChunkCache I'm not sure this is the way to go (even for video, but that seems to solve problems with big frames now).

If I'm not wrong, the chunks contain encoded data that is then passed to the decoder, so basing the size on the final decoded size does not make a lot of sense except maybe as an upper bound. And then chunk sizes are determined by the readers, not controlled from ChunkCache.

If I'm wrong maybe none of this makes sense.

Some measurements I've done with just a few local files, so probably not very representative:
- UHD video: 24883200 calculated cache size, the first chunk read is 1.2MB, with the rest being 100 and 400 KB. My machine is not powerful enough for 4K video, so these numbers may be very flawed.
- A random smallish mp4 video: 921600 calculated cache size, most chunks read are 2-4 KB, with the very first and a few more now and then being a bit more than 20KB.
- An mp3 audio file: 16384 calculated cache size, all chunks read were 626 or 627 bytes.
- A wav audio file: 8192 calculated cache size, all chunks read were 4K.

Now for the working of the cache itself: ChunkCache::SpaceLeft() returns true if not all (currently 10) entries are used and the memory available in the pool is greater than sizeof(chunk_buffer) + 2048. That's less than the typical chunk in all of the cases above expect for the mp3, which means that ReadNextChunk will be called, a new chunk_buffer allocated if needed, the next chunk read by the reader and then nixed because the buffer allocation fails.

With the numbers above, the wav case is affected and that's what's causing #16738. I'd say (but can't prove) that would also be the issue with 4K video. With the previous 3MB you'd need an average chunk size less than 300KB (in a perfect world without fragmentation with all those alloc/free and the management overhead).

I think the cache size should be managed by ChunkCache itself, but I don't have a safe, sensible and not wasteful solution, so I'll shut up as long as it works. The video part seems to work and waste a lot of memory for those two files, but the previous state would not work for the 4K video and would even be more wasteful for the mp4. The audio part wastes memory for the mp3 (the previous state would waste way more) and doesn't work for the wav. As noted in the commit message, increasing the magic 2000 would work, so that may be enough to keep going. Even making it 20000 would be a memory improvement over the previous hardcoded 3MB, but there's still the bad sensation that the calculation has nearly nothing to do with what we want to get.

Patch-set: 1
Reviewer: Gerrit User 1000223 <1000223@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-1