Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Please don't send this to beta, without at least increasing the audio magic number: #16738.
> 
> After some logging in ChunkCache I'm not sure this is the way to go (even for video, but that seems to solve problems with big frames now).
> 
> If I'm not wrong, the chunks contain encoded data that is then passed to the decoder, so basing the size on the final decoded size does not make a lot of sense except maybe as an upper bound. And then chunk sizes are determined by the readers, not controlled from ChunkCache.

[...]

> there's still the bad sensation that the calculation has nearly nothing to do with what we want to get.

You are correct, the computation here is only an upper bound (expecting that compressed frames will not be larger than uncompressed ones, and keeping a roughly similar size as before for lower resolution streams). It is indeed not perfect, but we don't really have a better guess, because we don't know anything about the encoded format being used. And, we can't really make this cache size dynamic, because it uses the "realtime malloc" allocator from the media kit. This allocator is designed to preallocate a large memory space, lock it in memory, and work from it. This avoids any of the data to be swapped out to disk, and make sure allocations also never need to swap out something else to get free RAM pages. It also is used only from one thread I think, so no risk of threads competing with each other for memory access. Basically, this allocator sacrifices everything (memory usage, thread safety, features) to be as fast as possible.

Given these constraints, we can do no more than approximations. Now we have an approximation that at least scales linearily with the video resolution and audio frame size, which I tihink is better than a fixed approximation. If there is a better proposal for how to size the buffer, we can incorporate it.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>