Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Should we keep support for disabling the chunk cache in that way? It's quite annoying if it ends up adding #if every two lines.
> 

Not my call to make. I didn't even know the option was there until I dived into that part and tried it for fun. I guess one can even just comment out the ChunkCache parts if they really needed for some testing or whatever and the option didn't exist, it's not that much.

> Is it possible to have a "dummy" implementation of the chunk cache that actually does no caching instead?

For the record, it currently (that is, without this patch) does not cache anything if DISABLE_CHUNK_CACHE is set to 1. It's just that the objects are created anyway, passing fExtractorWaitSem without the semaphore having been created, and later tried to release in MediaExtractor::Seek -> ChunkCache::MakeEmpty.

Patch-set: 1
Reviewer: Gerrit User 1000223 <1000223@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>