Update patch set 1
Patch Set 1:

> Again, why not fix SpaceLeft instead of finding the right "magic value"?

Any suggestion? From the comments thus far it would seem it's impossible without adding functionality in other places. You would have to know the size of the next chunk the reader may return and compare that to the biggest possible allocation (not the available memory in the pool, which I *guess* is what rtm_available gives). That or be able to give the reader an upper limit to the chunk size (which may or not be problematic later in the decoder). Or a way to tell the reader: hey that chunk you just gave me? Please keep it and return it when I ask again, I can't handle it now. But I'm no CS Eng, so there surely is something else obvious and easy.

Any of those would be better, yes, as there's no magic value guaranteed to work, for the very same reasons you can't know if you'll be able to handle the next chunk.

Patch-set: 1
Reviewer: Gerrit User 1000223 <1000223@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>