Update patch set 1
Patch Set 1:

Previous discussion (without a resolution as of this writing): https://review.haiku-os.org/c/haiku/+/3610

This is just a band-aid, not a solution. You may prefer to wait for a real solution or even disable the chunk cache completely.

Only tested with local files.

A lot (compared to the minimum needed) of wasted pinned memory for most playable files, but still not enough for some mp4 files (all MPEG4 ALS files I've tested want to read the whole file in one chunk) and maybe others.

There's still the theoretical possibility of not being able to play a video file that before could be played if: 3MB would be enough, the computed size is bigger and the system has just enough free memory to provide 3MB but not the computed size.

Patch-set: 1
Reviewer: Gerrit User 1000223 <1000223@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>