Update patch set 1
Patch Set 1:

> A lot (compared to the minimum needed) of wasted pinned memory for most playable files, but still not enough for some mp4 files (all MPEG4 ALS files I've tested want to read the whole file in one chunk) and maybe others.

It's a bit annoying to have to track this info in multiple reviews in Gerrit and in trac tickets (most of them merged or closed as duplicates).

Could you provide a summary of:
- The different file formats you used, and the resulting chunk sizes
- The size that we need the chunk cache to be (my understanding is that it's something kile constant1 + 10 * (constant2 + chunk size), is that correct?)

Also, complaining that things are "only a band-aid and not a solution" does not help a lot if there is no solution available. What do you suggest? How do thing behave with the chunk cache disabled? I think you tried this. Did it work? If so, why not remove this cache that can't be sized properly anyway? I expect that ffmpeg will already be doing its own caching anyway.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>