Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> Adding the check for B_METHOD_GET is easy. The bigger problem is the meaning of "bytesTotal" and "bytesReceived" for compressed transfers. This probably shouldn't be merged before investigating it.

If I read the code correctly, in the case of compressed transfers, bytesReceived and bytesTotal both apply to the received data before decoding/decompressing. They both match together, so the code as it is now is in fact correct.

The actual received file size will be larger (that will be tracked by bytesUnpacked as the data is decompressed). The total size isn't known in advance as the server may do the comnpression on the fly when processing the request. But that doesn't matter for this check: we can reliably detect the end of transfer using the compressed sizes.

Patch-set: 2