BTextView: Fix scrolling in text view
Prevent scrolling out of bounds horizontally. The algorithm to keep
cursor in text view bounds was taken from _PerformAutoScrolling()
(mouse scrolling method) and was fixed to account for the left
margin which is not always 0.

Fix left and top margin scrolling not always 0 in
_PerformAutoScrolling() as well, this fixes a similar bug that
was present when scrolling the text with the mouse instead.

Remove extraSpace variable, no longer needed.

This makes text views behave properly when overflow occurs i.e. when
you type text off the end of the text view. Instead of scrolling an
entire page over, it nudges the text over as you'd expect.

Replace xDiff and yDiff float variables with a single scrollTo BPoint
in ScrollToOffset(). This was also copied from _PerformAutoScrolling().

Remove floorf() calls, we know that we are dealing with integers.

Take top and bottom margins into account when scrolling with keyboard
in ScrollToOffset(). This allows you to scroll past the margins all the
way to the top and bottom using the up and down arrow keys (1 more px.)

Fixes #16642, #16476

Change-Id: I2d32d6039944d2dc3218ce4de71f2966cc98c866