BTextView: Fix scrolling in text view
Prevent scrolling out of bounds horizontally. Algorithm to keep
cursor in text view bounds was copied from _PerformAutoScrolling()
(mouse scrolling method) and was fixed to account for the left
margin which is not always 0.

Fix left and top margin scrolling not always 0 in
_PerformAutoScrolling() as well, this fixes a similar bug that
was present when scrolling the text with the mouse instead.

Remove extraSpace variable, no longer needed.

Replace xDiff and yDiff float variables with a single scrollTo BPoint
in ScrollToOffset(). This was also copied from _PerformAutoScrolling().

Remove a floorf() call, we know that we are dealing with integers.

Fixes #16642, #16476