Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> SATA trim is rightly disabled in the beta branch, so that is irrelevant for now. TRIM for NVMe is presently unimplemented.

It's implemented for ramdisks for some time already (I had found no problems while testing that, probably because I didn't create a partition table in a ramdisk). And the next commit adds support in the SD/mmc driver (which is how I found this problem while testing).

Maybe after fixing this it's time to test it again on SATA, and implement it on NVMe.

Rather than porting this to beta2, I'd say it's time to do a beta3 :)

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>