Update patch set 2
Patch Set 2:

> This makes me wonder if the HID parser should be moved to input_server:

I think the only major argument to keep it is in kernel space is largely because of USB keyboard support in KDL. So, we would need to find some way to inform KDL of USB keyboard HID report layout and then some minimum amount of code to keep it functioning. But yes, if we can solve that problem, moving HID parsing to input_server sounds like a good idea.

Patch-set: 2
CC: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>