Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> When resizing StyledEdit window with vertically scrolled text, after first resize window become empty and scroll bars become disabled. After second resize text and scroll bars are recovered, but text are scrolled to beginning.

I can't seem to reproduce this. Can you give me more detailed steps on how you were able to trigger this bug? Also does the bug happen without this patch as well? I opened a file long enough that the vertical scroll bar is enabled in StyledEdit. I scrolled down a bit vertically to move from the top of the document, then I clicked on the window to change my cursor position, then I resized the window by dragging the bottom-right corner of the window. The scroll position changed a bit as the word-wrapping updated but the scroll bar did not become disabled and the text was not scrolled to the beginning.

The good news about the bug I found above is that it happens even without this patch. The bad news is that it still happens with this patch too. So there is a bug somewhere else.

Patch-set: 2
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>