BTextView: Fix drawing issues when typing off the edge
When word-wrap is turned on for a multi-line text view and you
insert, backspace or delete the position of text rect needs to
change as you type off the end of the BTextView area. Call
FrameResized() in ScrollToOffset() to fix the problem. This can
also happen as a result of word-wrap setting changing so
reposition the text rect and redraw in that case too.

Use Window()->DisableUpdates() and Window()->EnableUpdates() and
remove a call to Window()->UpdateIfNeeded() to prevent the text
from flashing as it is drawn then the text rect position is moved
and draws again.