Update patch set 1
Patch Set 1:

unichar was defined in SupportDefs in BeOS:
https://github.com/jscipione/r5headers/blob/master/headers/be/support/SupportDefs.h#L59

So, can we really move or remove it? (ok, we only provide binary compatibility, so it should be fine)

Should we make sure that our POSIX headers don't end up including Be headers like SupportDefs.h? Why did it end up being included in irssi? Is it through standard POSIX headers, or is it because they use the Be API directly for some things?

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>