Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-1
> 1. Then typing long name and then erasing, text field do not become smaller, it only grows.

I believe the problem you described is caused by https://git.haiku-os.org/haiku/tree/src/kits/interface/TextView.cpp#n3887

When comparing the behavior of BTextView on BeOS R5 I found that the width of fTextRect (the area of the text inside BTextView) is always set to the length of the string, while Haiku since my patch the width is at least as wide as the parent BTextView container and it only grows from there, it doesn't ever shrink. This works fine most of the time because the width of the container BTextView usually doesn't change and so you can't see that there is a wider fTextRect inside it. However in Tracker Edit name, because the width BTextView is unconstrained, you can see that it is made wider and the width of the BTextView container does change.

I could fairly easily alter this part of the code to behave like BeOS R5, i.e. make the text rect width always equal to the width of the containing string. The problem with this approach is that currently we never grow a 0 width fTextRect, which is what you'll get with a blank BTextView. The reason for this is performance, if you start at 0 width fTextRect then _RecalculateLineBreaks() needs to go through each line of text to grow fTextRect more and more until the text all fits. This was a problem in AboutSystem for example because there's many lines of text that need to have their line breaks calculated for.

Previous to my patch we set a minimum 10px width on fTextRect and that made _RecalculateLineBreaks() work, albeit slowly. So, the question is, how can we get the BeOS R5 behavior without paying the performance penalty needed to grow and/or shrink the fTextRect in the BTextView?

Patch-set: 3
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>